Moved 'test_tensorrt_preprocess.py' and 'preprocess', updated 'remove_tmp_file'.
